requirements.txt
instagrapi==2.0.0
python-telegram-bot==21.6
python-dotenv==1.0.1
# Optional: fallback tools
instaloader==4.11
urllib3==1.26.18
Pillow>=8.1.1
gunicorn==21.2.0
Flask==3.0.3
